Fix for https://wso2.org/jira/browse/IOTS-155 & https://wso2.org/jira/browse/IOTS-164
