--- components/device-mgt/org.wso2.carbon.device.mgt.ui/src/main/resources/jaggeryapps/devicemgt/app/units/cdmf.unit.policy.listing/listing.json
@@ -1,3 +1,0 @@
-{
-	"version": "1.0.0"
-}
